Address Properties moved to addresses controller
