+++ config/routes.js
@@ -33,11 +33,10 @@ module.exports = function(app) {
   app.get(apiPrefix + '/addr/:addr/utxo', addresses.utxo);
 
   // Address property routes
-  var addressProperties = require('../app/controllers/addressProperties.js');
-  app.get(apiPrefix + '/addr/:addr/balance', addressProperties.balance);
-  app.get(apiPrefix + '/addr/:addr/totalReceived', addressProperties.totalReceived);
-  app.get(apiPrefix + '/addr/:addr/totalSent', addressProperties.totalSent);
-  app.get(apiPrefix + '/addr/:addr/unconfirmedBalance', addressProperties.unconfirmedBalance);
+  app.get(apiPrefix + '/addr/:addr/balance', addresses.balance);
+  app.get(apiPrefix + '/addr/:addr/totalReceived', addresses.totalReceived);
+  app.get(apiPrefix + '/addr/:addr/totalSent', addresses.totalSent);
+  app.get(apiPrefix + '/addr/:addr/unconfirmedBalance', addresses.unconfirmedBalance);
 
   // Status route
   var st = require('../app/controllers/status');
